pyproject.toml
[project]
name = "cv-chat-agent"
version = "0.1.0"
description = "Simple CV chat agent using LLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.61.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
]
